+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix `ContainerDetails::config` field deserialization
 - Fix `LogConfig::config` field deserialization
 - Fix network creation [#11](https://github.com/vv9k/docker-api-rs/pull/11)
+- Add missing `NetworkInfo::created` field
 
 # 0.7.0
 - Make `PidsStats`, `Descriptor` and `DistributionInspectInfo` struct fields publicly accessible.